Use idMaker for dataadaptercache
