+++ packages/core/data_adapters/BaseAdapter.ts
@@ -2,7 +2,7 @@ import { Observable, merge } from 'rxjs'
 import { takeUntil } from 'rxjs/operators'
 import { isStateTreeNode, getSnapshot } from 'mobx-state-tree'
 import { ObservableCreate } from '../util/rxjs'
-import { checkAbortSignal, hashCode, observeAbortSignal } from '../util'
+import { checkAbortSignal, observeAbortSignal } from '../util'
 import { Feature } from '../util/simpleFeature'
 import {
   AnyConfigurationModel,
@@ -12,6 +12,7 @@ import { getSubAdapterType } from './dataAdapterCache'
 import { Region, NoAssemblyRegion } from '../util/types'
 import { blankStats, rectifyStats, scoresToStats } from '../util/stats'
 import BaseResult from '../TextSearch/BaseResults'
+import idMaker from '../util/idMaker'
 
 export interface BaseOptions {
   signal?: AbortSignal
@@ -48,26 +49,6 @@ export type AnyDataAdapter =
   | BaseTextSearchAdapter
   | RegionsAdapter
   | SequenceAdapter
-
-// generates a short "id fingerprint" from the config passed to the base
-// feature adapter by recursively enumerating props, but if config is too big
-// does not process entire config (FromConfigAdapter for example can be large)
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function idMaker(args: any, id = '') {
-  const keys = Object.keys(args)
-  for (let i = 0; i < keys.length; i++) {
-    const key = keys[i]
-    if (id.length > 5000) {
-      break
-    }
-    if (typeof args[key] === 'object' && args[key]) {
-      id += idMaker(args[key], id)
-    } else {
-      id += `${key}-${args[key]};`
-    }
-  }
-  return hashCode(id)
-}
 
 export abstract class BaseAdapter {
   public id: string
